Hudmanager.Updateを統一
